pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "longevity-lens"
version = "0.1.0"
description = "Agentic system that scrapes, standardises, and reasons over scientific evidence for aging interventions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Data models & settings
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # HTTP client
    "httpx>=0.27.0",
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    # LLM
    "openai>=1.0",
    "google-genai>=1.0.0",
    # Storage
    "aiosqlite>=0.20.0",
    # Search
    "tavily-python>=0.5.0",
    # Google Trends
    "pytrends>=4.9.0",
    # Frontend
    "streamlit>=1.40.0",
    # Logging
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.34.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
